Simpler way for HTTP
