+++ src/Arcus.Shared/Services/BaconService.cs
@@ -1,11 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Arcus.Observability.Correlation;
-using Arcus.Observability.Telemetry.Core;
-using Arcus.POC.Observability.Telemetry.Serilog.Sinks.ApplicationInsights.Extensions;
 using Arcus.Shared.Services.Interfaces;
 using GuardNet;
 using Microsoft.Extensions.Configuration;
@@ -16,12 +13,13 @@ namespace Arcus.Shared.Services
 {
     public class BaconService : IBaconService
     {
+        private readonly IConfiguration _configuration;
         private readonly ICorrelationInfoAccessor _correlationInfoAccessor;
         private readonly IHttpClientFactory _httpClientFactory;
-        private readonly IConfiguration _configuration;
         private readonly ILogger<BaconService> _logger;
 
-        public BaconService(IHttpClientFactory httpClientFactory, ICorrelationInfoAccessor correlationInfoAccessor, IConfiguration configuration, ILogger<BaconService> logger)
+        public BaconService(IHttpClientFactory httpClientFactory, ICorrelationInfoAccessor correlationInfoAccessor,
+            IConfiguration configuration, ILogger<BaconService> logger)
         {
             Guard.NotNull(correlationInfoAccessor, nameof(correlationInfoAccessor));
             Guard.NotNull(httpClientFactory, nameof(httpClientFactory));
@@ -41,11 +39,8 @@ namespace Arcus.Shared.Services
             var request = new HttpRequestMessage(HttpMethod.Get, $"http://{url}/api/v1/bacon");
 
             var response = await SendHttpRequestAsync("Get Bacon", request);
-            if (response.IsSuccessStatusCode == false)
-            {
-                throw new Exception("Unable to get bacon");
-            }
-            
+            if (response.IsSuccessStatusCode == false) throw new Exception("Unable to get bacon");
+
             var rawResponse = await response.Content.ReadAsStringAsync();
             return JsonConvert.DeserializeObject<List<string>>(rawResponse);
         }
@@ -53,25 +48,11 @@ namespace Arcus.Shared.Services
         private async Task<HttpResponseMessage> SendHttpRequestAsync(string operationName, HttpRequestMessage request)
         {
             var httpClient = _httpClientFactory.CreateClient();
+            var response = await httpClient.SendAndTrackDependencyAsync(operationName, request, _correlationInfoAccessor, _logger);
 
-            using(var httpDependencyMeasurement = DependencyMeasurement.Start(operationName))
-            {
-                // TODO: Verify
-                var newDependencyId = Guid.NewGuid().ToString();
-                var correlationInfo = _correlationInfoAccessor.GetCorrelationInfo();
-                var upstreamOperationParentId = $"|{correlationInfo?.OperationId}.{newDependencyId}";
-                request.Headers.Add("Request-Id", upstreamOperationParentId);
-                request.Headers.Add("X-Transaction-ID", correlationInfo?.TransactionId);
+            _logger.LogInformation("Calling bacon API completed with status:" + response.StatusCode);
 
-                // TODO: Check HTTP pipeline for hooks
-                // https://thomaslevesque.com/2016/12/08/fun-with-the-httpclient-pipeline/
-                var response = await httpClient.SendAsync(request);
-                
-                _logger.LogInformation("Calling bacon API completed with status:" + response.StatusCode);
-                _logger.LogExtendedHttpDependency(request, response.StatusCode, httpDependencyMeasurement, dependencyId: upstreamOperationParentId);
-
-                return response;
-            }
+            return response;
         }
     }
 }
